requirements.txt
# Core
discord.py
git+https://github.com/Rapptz/discord-ext-menus@master
asyncpg # Async PostgreSQL

pendulum # Advanced Date/Time
parsedatetime # Read time intervals from user in
jishaku # Dev tools
psutil # System info
fuzzywuzzy # Fuzzy Searching
python-Levenshtein # ^
d20 # Dice Roller